Working Nodes

+++ Scripts/Globals.gd
@@ -9,7 +9,7 @@ func _input(event):
 	if event is InputEventMouseButton and event.is_pressed():
 		match event.button_index:
 			BUTTON_RIGHT:
-				held_line.delete()
+				if held_line: held_line.delete()
 
 func _process(delta):
 	if held_line and is_instance_valid(held_line):
